feat: restructure ChatClient and useChat for improved organization and functionality
